Update README.md
Starting a visdom server is necessary in order to train a model, however it is not mentioned at all in the readme and requires visiting the issues or FAQ page to learn what the `error connecting to visdom server` is all about. Suggest just putting it as a step before `train a model`. Thanks!